修复了一些因为长整型long带来的问题

+++ src/main/java/com/st/newtest/stGame/Service/WebActionService.java
@@ -5,7 +5,7 @@ import com.st.newtest.stGame.Entity.WebAction;
 import java.util.List;
 
 public interface WebActionService {
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(WebAction record);
